% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulators.R
\name{append_popsize}
\alias{append_popsize}
\title{Append the population sizes for each respective population to a data frame.}
\usage{
append_popsize(df, pop_df, left_popref = pop, right_popref = pop)
}
\arguments{
\item{df}{Data frame lacking population size column.}

\item{pop_df}{Data frame containing population size information with key
pertaining to the population.}

\item{left_popref}{The name of the population reference column of `df`.}

\item{right_popref}{The name of the population reference column of `pop_df`.}
}
\value{
The original data frame with a population size column.
}
\description{
Use this function when a data frame has a reference to a population (ex:
  population_id) but is lacking a population size for each population.
}
