% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulators.R
\name{create_population_pq}
\alias{create_population_pq}
\title{Create a frame of of p and q for multiple populations}
\usage{
create_population_pq(data, pop_freq_vars)
}
\arguments{
\item{data}{A data frame contain 1+ columns of allele frequencies.}

\item{pop_freq_vars}{A vector of unquoted allele frequency vars.}
}
\value{
A data frame with new columns of `pop`, `p` and `q`.
}
\description{
Takes in a vector of columns that reference mutation population frequencies
  and converts them to p and q values where q is the original frequency at
  population i and p = 1 - q. The resulting frame also indexes the
  populations and pivots the table to long format such that each row is a
  reference to a single population under certain parameters.
}
