% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulators.R
\name{expand_genome_to_template}
\alias{expand_genome_to_template}
\title{Join template genome position to a single parameter
of mutations}
\usage{
expand_genome_to_template(mutation_df, template, generation_var, param_vars)
}
\arguments{
\item{mutation_df}{A data frame of mutations for a given
parameter set}

\item{template}{A data frame of all the positions on the
genome corresponding to the mutations}

\item{generation_var}{The variable in the `mutation_df`
corresponding to the the output generation variable}

\item{param_vars}{A vector of parameter variable names in
the `mutation_df` frame}
}
\value{
An expanded mutation data frame with empty (neutral)
 position rows
}
\description{
The default output mutations doesn't include a value for
 neutral positions as a way to save resources when simulations
 are running. This function joins on the template genome
 (essentially a template for the positions on the genome) to
 the mutation data so that each position per each generation
 has a row. This may or may not be the appropriate method
 depending on how you plan to use the data next.
}
